Fully Functional Tower placement mechanics. :)
Once the Tower UI button is clicked, the player can place a tower on a grass tile. Will create a marker showing each tile that a tower can be placed on, will throw an error if a non-grass tile is clicked, and will properly reset once a tower is placed.